fix: renamed pages to screens
