+++ lib/app/auth_gate.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pet_haven/pages/login.dart';
-import 'package:pet_haven/pages/main_screen.dart';
+import 'package:pet_haven/screens/login.dart';
+import 'package:pet_haven/screens/main_screen.dart';
 import 'package:provider/provider.dart';
 import '../data/auth_manager.dart';
