+++ lib/screens/main_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:pet_haven/components/app_bottom_navigation.dart';
-import 'package:pet_haven/pages/cart.dart';
-import 'package:pet_haven/pages/home.dart';
-import 'package:pet_haven/pages/profile.dart';
-import 'package:pet_haven/pages/shop.dart';
+import 'package:pet_haven/screens/cart.dart';
+import 'package:pet_haven/screens/home.dart';
+import 'package:pet_haven/screens/profile.dart';
+import 'package:pet_haven/screens/shop.dart';
 
 /// The main container widget for the Pet Haven app.
 ///
