+++ lib/screens/profile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pet_haven/components/custom_app_bar.dart';
 import 'package:pet_haven/components/profile_titles.dart';
-import 'package:pet_haven/pages/about_app.dart';
+import 'package:pet_haven/screens/about_app.dart';
 import 'package:pet_haven/theme/color.dart';
 
 class Profile extends StatelessWidget {
