+++ lib/screens/shop.dart
@@ -5,7 +5,7 @@ import 'package:pet_haven/components/input_field.dart';
 import 'package:pet_haven/components/custom_card.dart';
 import 'package:pet_haven/data/product.dart';
 import 'package:pet_haven/data/product_repository.dart';
-import 'package:pet_haven/pages/product_detail_page.dart';
+import 'package:pet_haven/screens/product_detail_page.dart';
 
 class Shop extends StatefulWidget {
   const Shop({super.key});
